# doc/api.yml
openapi: 3.0.3
info:
  version: 'latest'
  title: HEI Admin API
  description: |
    _Programmatically connect to a computer programming [school](https://hei.school)._
    After [joining us](mailto:contact@mail.hei.school),
    you can get an identification token from
    our [dev](https://dev-hei-admin.auth.eu-west-3.amazoncognito.com/oauth2/authorize?client_id=5s8cg50doahmu855rlc8fr6qmp&response_type=token&scope=email+openid&redirect_uri=https%3A%2F%2Fapi-dev.hei.school%2Fwhoami)
    or [prod](https://prod-hei-admin.auth.eu-west-3.amazoncognito.com/oauth2/authorize?client_id=i8bg538jpfu6mqmqb61m26trd&response_type=token&scope=email+openid&redirect_uri=https%3A%2F%2Fapi-prod.hei.school%2Fwhoami)
    authentication service.
    Then, start playing with our system!

    The implementation of our API is [publicly disclosed](https://github.com/hei-school/hei-admin-api).
    You are welcome to try and compromise it. Happy hacking!
servers:
  - url: 'https://api-dev.hei.school'
  - url: 'https://api-prod.hei.school'
security:
  - BearerAuth: [ ]
paths:
...
  '/teachers/{id}':
    get:
      tags:
        - Users
      summary: Get teacher by identifier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      operationId: getTeacherById
      responses:
        '200':
          description: The identified teacher
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Teacher'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
    put:
      tags:
        - Users
      summary: Update teacher
      operationId: updateTeacher
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: Teacher to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrupdateTeacher'
      responses:
        '200':
          description: updated teacher.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Teacher'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
  /teachers:
    get:
      tags:
        - Users
      summary: Get all teachers
      operationId: getTeachers
      parameters:
        - name: page
          in: query
          schema:
            $ref: '#/components/schemas/Page'
        - name: page_size
          in: query
          schema:
            $ref: '#/components/schemas/PageSize'
        - name: ref
          in: query
          description: "Filter teachers by ref, case is ignored"
          schema:
            type: string
        - name: first_name
          in: query
          description: "Filter teachers by first name, case is ignored"
          schema:
            type: string
        - name: last_name
          in: query
          description: "Filter teachers by last name, case is ignored"
          schema:
            type: string
        - name: status
          required: false
          in: query
          description: Filter teachers by status value return all by default
          schema:
            $ref: '#/components/schemas/EnableStatus'
        - name: sex
          required: false
          in: query
          description: Filter teachers by sex value return all by default
          schema:
            $ref: '#/components/schemas/Sex'
      responses:
        '200':
          description: List of teachers, ordered by ref.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Teacher'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
    put:
      tags:
        - Users
      summary: Create new teachers or update existing teachers
      description: Update teachers when `id` are provided, create them otherwise.
      operationId: createOrUpdateTeachers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CrupdateTeacher'
      responses:
        '200':
          description: The created or updated teachers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Teacher'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
...
  '/teachers/{teacher_id}/picture/raw':
    post:
      tags:
        - Users
      summary: Upload teacher profile picture
      operationId: uploadTeacherProfilePicture
      parameters:
        - name: teacher_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UploadFile"
      responses:
        '200':
          description: Teacher updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Teacher'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
...
  '/promotions':
    get:
      tags:
        - Promotions
      operationId: getPromotions
      summary: Get all promotions
      parameters:
        - name: page
          in: query
          schema:
            $ref: '#/components/schemas/Page'
          description: Set value to 1 by default if null is provided
          required: false
        - name: page_size
          in: query
          description: Set value to 15 by default if null is provided
          schema:
            $ref: '#/components/schemas/PageSize'
          required: false
        - name: name
          in: query
          description: Filter promotion by name
          schema:
            type: string
          required: false
        - name: ref
          in: query
          description: Filter promotion by ref
          schema:
            type: string
          required: false
        - name: group_ref
          in: query
          description: Filter promotion by group ref
          schema:
            type: string
          required: false
      responses:
        '200':
          description: Promotions filtered with criteria if provided
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Promotion'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
    put:
      tags:
        - Promotions
      operationId: crupdatePromotion
      summary: Update promotion if `id` provided, create otherwise
      requestBody:
        description: Create or update promotion
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrupdatePromotion'

      responses:
        '200':
          description: The promotion created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Promotion'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
  '/promotions/{id}':
    get:
      tags:
        - Promotions
      operationId: getPromotionById
      summary: Get promotion by the given id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The promotion queried
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Promotion'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
  '/promotions/{id}/students':
    get:
      tags:
        - Promotions
      operationId: getStudentsByPromotion
      summary: Get students in the promotion
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of students in a xlsx file
          content:
            application/xlsx:
              schema:
                type: string
                format: byte
                example: XLSX VERSION
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
  '/promotions/{id}/groups':
    put:
      tags:
        - Promotions
      operationId: updatePromotionGroups
      summary: Add or remove promotion's groups
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: List of group ids to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePromotionSGroup'
      responses:
        '200':
          description: The promotion created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Promotion'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  responses:
    '400':
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestException'
    '403':
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotAuthorizedException'
    '404':
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResourceNotFoundException'
    '429':
      description: Too many requests to the API
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TooManyRequestsException'
    '500':
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalServerException'
  schemas:
...
    UserIdentifier:
      type: object
      properties:
        id:
          type: string
        ref:
          type: string
          example: STD000001
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        nic:
          type: string
...
    CrupdateUser:
      allOf:
        - $ref: '#/components/schemas/UserIdentifier'
        - type: object
          properties:
            sex:
              $ref: '#/components/schemas/Sex'
            birth_date:
              type: string
              format: date
            birth_place:
              type: string
            address:
              type: string
            phone:
              type: string
            entrance_datetime:
              type: string
              format: date-time
            coordinates:
              $ref: '#/components/schemas/Coordinates'
            high_school_origin:
              type: string
            status:
              $ref: '#/components/schemas/EnableStatus'
...
    Sex:
      type: string
      enum:
        - M
        - F
    EnableStatus:
      type: string
      enum:
        - ENABLED
        - DISABLED
        - SUSPENDED
...
    CrupdateTeacher:
      allOf:
        - $ref: '#/components/schemas/CrupdateUser'
...
    Teacher:
      allOf:
        - $ref: '#/components/schemas/CrupdateUser'
      properties:
        profile_picture:
          type: string
...
    GroupIdentifier:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          example: The group number one
        ref:
          type: string
          example: G1
...
    Coordinates:
      type: object
      properties:
        longitude:
          type: number
          format: double
        latitude:
          type: number
          format: double
...
    CrupdatePromotion:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        ref:
          type: string
    UpdatePromotionSGroup:
      type: object
      properties:
        type:
          type: string
          enum:
            - ADD
            - REMOVE
        group_ids:
          type: array
          items:
            type: string
    Promotion:
      type: object
      allOf:
        - $ref: '#/components/schemas/CrupdatePromotion'
      properties:
        creation_datetime:
          type: string
          format: date-time
        groups:
          type: array
          items:
            $ref: '#/components/schemas/GroupIdentifier'
...
    UploadFile:
      type: object
      properties:
        file_to_upload:
          type: string
          format: binary
          description: User profile picture
...
    Page:
      type: integer
      description: Start at `1`
    PageSize:
      type: integer
      description: Cannot exceed `500`
      default: 15
...
    Exception:
      type: object
      properties:
        type:
          type: string
        message:
          type: string
    BadRequestException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: BadRequestException
        message: Bad request
    NotAuthorizedException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: NotAuthorizedException
        message: Not authorized
    ResourceNotFoundException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: ResourceNotFoundException
        message: Resource of type <T> identified by <I> not found
    TooManyRequestsException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: TooManyRequestsException
        message: Too many requests
    InternalServerException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: InternalServerException
        message: Unexpected error